Fix for IE8 unbound buttons
Hopscotch doesn't work in IE8. Hopscotch 0.2.0 broke it as in `_handleBubbleClick` `evt` argument will be undefined as it's not passed by IE8. Instead you have to grab the event from `window.event` object and find `currentTarget` as `evt.srcElement`.

It probably requires some testing but seems to work for me.