Fix scroll animation callback called twice
jQuery calls the animation callback for each of the matched elements in the selector. Hopscotch was selecting both the `html` element and the `body` element for adjusting the `scrollTop` value, leading to the callback being called twice.

Adjusting `scrollTop` on the `html` element seems questionable anyway, so just removing it fixes the issue.