[Chore] Refactor notice; Bump year
Add a notice to README.md and CONTRIBUTING.md that clarifies the current state of Hopscotch development. While we're still happy to accept bug reports, feature requests, and PRs, I want to be clear about the fact that they may not get addressed for a while.

Also, bump year in Gruntfile.